refactor: carry mocks to upper level
